requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
args==0.1.0
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.3
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
cmake==3.31.4
contextlib2==21.6.0
datasets==3.3.2
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
hpsv2==1.2.0
huggingface-hub==0.24.7
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
lit==18.1.8
Markdown==3.7
MarkupSafe==3.0.2
ml_collections==0.1.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
openai-clip==1.0.1
packaging==24.2
pandas==2.2.3
peft==0.8.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
timm==1.0.15
tokenizers==0.15.2
tomli==2.2.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.7
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
xformers==0.0.23.post1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
inflect
